LRCI-7783 Add WorkspaceGitRepositoryTest#testSetUp()
